Refactoring - Shifted Polygon Playground to selfie.md - Shifted Sponsorship page to .md - Moved more icons and patterns to static folder
