+++ .prettierrc
@@ -3,10 +3,16 @@
 	"singleQuote": true,
 	"trailingComma": "none",
 	"printWidth": 100,
-	"plugins": ["prettier-plugin-svelte", "prettier-plugin-tailwindcss"],
+	"plugins": [
+		"prettier-plugin-svelte",
+		"prettier-plugin-tailwindcss"
+	],
 	"overrides": [
 		{
-			"files": "*.svelte",
+			"files": [
+				"*.svx",
+				"*.svelte"
+			],
 			"options": {
 				"parser": "svelte"
 			}
