+++ content/tools/selfie.md
@@ -1,0 +1,13 @@
+---
+title: Create your VizChitra Data Selfie
+description: An interactive playground to create your own unique VizChitra Polygon Data Selfie
+---
+
+<script>
+  import {FullBleed} from '$lib/components/layout'
+  import {PolygonPlayground} from '$lib/components/sections'
+</script>
+
+<FullBleed>
+   <PolygonPlayground/>
+</FullBleed>
