+++ src/lib/components/interface/index.ts
@@ -6,3 +6,4 @@ export { default as ButtonBar } from './ButtonBar.svelte';
 export { default as Dropdown } from './Dropdown.svelte';
 export { default as SpeakerDeck } from './SpeakerDeck.svelte';
 export { default as ColumnChart } from './ColumnChart.svelte';
+export { default as Card } from './Card.svelte';
