+++ src/lib/components/sections/playground/PolygonGenerator.svelte
