+++ src/lib/components/sections/playground/SliderInput.svelte
