+++ src/lib/components/sections/playground/index.ts
@@ -2,3 +2,4 @@ export { default as ImageUpload } from './ImageUpload.svelte';
 export { default as SliderInput } from './SliderInput.svelte';
 export { default as PolygonGenerator } from './PolygonGenerator.svelte';
 export { default as SelectInput } from './SelectInput.svelte';
+export { default as PolygonPlayground } from './PolygonPlayground.svelte';
