--- src/lib/components/sections/sponsors/index.ts
@@ -1,1 +1,0 @@
-export { default as SponsorCards } from './SponsorCards.svelte';
