+++ src/lib/components/sections/sponsorship/index.ts
@@ -1,1 +1,5 @@
 export { default as SponsorshipContactForm } from './SponsorshipContactForm.svelte';
+export { default as AttendeeCharts } from './AttendeeCharts.svelte';
+export { default as SponsorCard } from './SponsorCard.svelte';
+export { default as AttendeeMarquee } from './AttendeeMarquee.svelte';
+export { default as SponsorButton } from './SponsorButton.svelte';
