+++ src/lib/components/typography/index.ts
@@ -1,2 +1,4 @@
 export { default as Slanted } from './Slanted.svelte';
 export { default as VizChitraLogoType } from './VizChitraLogoType.svelte';
+export { default as Heading } from './Heading.svelte';
+export { default as SubHeading } from './SubHeading.svelte';
