+++ src/lib/config/redirects.js
@@ -4,22 +4,11 @@
 export const redirects = {
 	friends: 'https://chat.whatsapp.com/CcptoGEeCqJEkiETwAzpbq',
 	recap: '/',
-	conference: 'https://hasgeek.com/VizChitra/2025',
-	cfp: 'https://hasgeek.com/VizChitra/2025/sub',
-	schedule: 'https://hasgeek.com/VizChitra/2025/schedule',
-	feedback: 'https://forms.gle/ai8npdr4GKHDHkZw5',
-	tickets: 'https://hasgeek.com/VizChitra/2025/#tickets',
-	flash: 'https://forms.gle/c62J2Pttascvy37W9',
-	workfeedback: 'https://forms.gle/qw5xB57QBh1h6P4Q9',
-	merchandise: 'https://hasgeek.com/VizChitra/merchandise/',
+	'polygon-playground': '/tools/selfie',
+	'sponsorship': '2025/sponsorship',
 	map: 'https://maps.app.goo.gl/Xa6N8q7s5S2Tm1559',
-	'sponsor-feedback': 'https://forms.gle/4TBRt6zTuMdMMiyKA',
 	guide: 'https://drive.usercontent.google.com/download?id=1g3e1RkHpv7pW9SZCR4m9FME8rKJ3lSo_&authuser=0&acrobatPromotionSource=GoogleDriveListView',
 	editorial: 'https://docs.google.com/presentation/d/1ODffg_I407aptk8zBzrzEXS4i-GiCCVExp2t2bQKZ7E/',
-	virtual1: 'https://www.youtube.com/watch?v=MJ8OT6J53o4',
-	virtual2: 'https://youtube.com/live/-g2wH27Qu58?feature=share',
-	slido1: 'https://app.sli.do/event/xf42twibEgDLDifMKdxDwW',
-	slido2: 'https://app.sli.do/event/5aJYogWhdFTVcg7No9gH8f',
 	2025: 'https://vizchitra.com',
 	'meet-the-team': "team",
 	'code-of-conduct': "conduct"
