--- src/lib/types/csv.d.ts
@@ -1,4 +1,0 @@
-declare module '*.csv' {
-	const content: any;
-	export default content;
-}
