+++ src/routes/[...slug]/+page.server.ts
@@ -3,21 +3,37 @@ import type { PageServerLoad } from './$types';
 import yaml from 'js-yaml';
 
 // Import all markdown files as raw strings so we can extract frontmatter for prerender entries.
-const contentModules = import.meta.glob('../../../content/*.md', {
+const contentModules = import.meta.glob('../../../content/**/*.md', {
 	query: '?raw',
 	import: 'default'
 });
 
 const contentMap: Record<string, string> = {};
 for (const path in contentModules) {
-	const fileName = path.split('/').pop()?.replace('.md', '') || '';
-	contentMap[fileName] = path;
+	const slug = path.replace('../../../content/', '').replace('.md', '');
+	contentMap[slug] = path;
 }
 
 export const prerender = true;
 
 export const entries = async () => {
-	return Object.keys(contentMap).map((slug) => ({ slug }));
+	const slugs = [];
+	for (const slug of Object.keys(contentMap)) {
+		const contentPath = contentMap[slug];
+		const content = (await contentModules[contentPath]()) as string;
+
+		// Extract frontmatter to check for draft status
+		const frontmatterRegex = /^---\n([\s\S]*?)\n---\n/;
+		const match = content.match(frontmatterRegex);
+		if (match) {
+			const frontmatter = (yaml.load(match[1]) || {}) as any;
+			if (import.meta.env.PROD && frontmatter.draft) {
+				continue;
+			}
+		}
+		slugs.push({ slug });
+	}
+	return slugs;
 };
 
 export const load: PageServerLoad = async ({ params }) => {
@@ -37,6 +53,10 @@ export const load: PageServerLoad = async ({ params }) => {
 		let frontmatter = {};
 		if (match) {
 			frontmatter = yaml.load(match[1]) || {};
+		}
+
+		if (import.meta.env.PROD && (frontmatter as any).draft) {
+			throw error(404, `Content page "${slug}" not found`);
 		}
 
 		return {
