--- src/routes/polygon-playground/+page.ts
@@ -1,1 +1,0 @@
-export const prerender = true;
