--- src/routes/sponsorship/+page.js
@@ -1,1 +1,0 @@
-export const prerender = true;
